requirements.txt
annotated-types==0.7.0
astroid==3.3.8
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dill==0.3.9
Flask==3.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.3
python-dotenv==1.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
